requirements.txt
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.6
matplotlib==3.7.2
fpdf==1.7.2
schedule==1.2.0
